+++ CHANGELOG.md
@@ -6,6 +6,18 @@ TENER EN CUENTA: (DESPS SE BORRA)
 - cuando se cambia tile, max 7, sino Exception
 - comodin no poner mas de 1 letra, exception si tengo rr 
 
+## [0.2.8] 2023-11-02
+
+### Added
+
+
+- add method iterate word 
+
+### Fixed
+
+
+- fixed place word
+
 ## [0.2.8] 2023-11-01
 
 ### Added